Take: Move the ValuesSpan class to the header
